added authorization to rule resources

+++ server/src/main/java/org/kyantra/resources/SnsRuleResource.java
@@ -7,11 +7,9 @@ import com.amazonaws.services.sns.model.CreateTopicResult;
 import io.swagger.annotations.Api;
 import org.kyantra.aws.RuleHelper;
 import org.kyantra.aws.SnsHelper;
-import org.kyantra.beans.RoleEnum;
-import org.kyantra.beans.RuleBean;
-import org.kyantra.beans.SnsBean;
-import org.kyantra.beans.SnsSubscriptionBean;
+import org.kyantra.beans.*;
 import org.kyantra.dao.*;
+import org.kyantra.helper.AuthorizationHelper;
 import org.kyantra.interfaces.Secure;
 import org.kyantra.interfaces.Session;
 import org.kyantra.services.ValidatorService;
@@ -31,12 +29,13 @@ public class SnsRuleResource extends BaseResource {
     /* TODO: Atomicity with AWS and DB
      * For transactions, like if DB transactions don't commit properly rollback the AWS operations too*/
 
+    // TODO: 5/25/18 Change id pathparam to formparam
     @POST
-    @Session
     @Path("/create/{id}")
+    @Session
+    @Secure(roles = {RoleEnum.WRITE, RoleEnum.ALL}, subjectType = "rule", subjectField = "parentId")
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.APPLICATION_JSON)
-    @Secure(roles = {RoleEnum.ALL}, subjectType = "rule", subjectField = "parentId")
     public String create(@PathParam("id") Integer parentThingId,
                          @FormParam("name") String name,
                          @FormParam("description") String description,
@@ -45,7 +44,7 @@ public class SnsRuleResource extends BaseResource {
                          @FormParam("subject") String subject,
                          @FormParam("message") String message,
                          @FormParam("interval") Integer interval,
-                         @FormParam("sns_topic") String snsTopic) {
+                         @FormParam("sns_topic") String snsTopic) throws AccessDeniedException {
         /*
          * Steps:
          * 1. create SnsBean
@@ -56,51 +55,121 @@ public class SnsRuleResource extends BaseResource {
          * 6. link rule and SNS in DB
          * */
 
-        // create SnsBean
-        SnsBean snsBean = new SnsBean();
-        snsBean.setTopic(snsTopic);
-        // set parameters or revert to default
-        snsBean.setSubject(subject);
-        snsBean.setMessage(message);
-        snsBean.setInterval(interval);
+        ThingBean targetThing = ThingDAO.getInstance().get(parentThingId);
+        UserBean user = (UserBean) getSecurityContext().getUserPrincipal();
+
+        if (AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            // create SnsBean
+            SnsBean snsBean = new SnsBean();
+            snsBean.setTopic(snsTopic);
+            // set parameters or revert to default
+            snsBean.setSubject(subject);
+            snsBean.setMessage(message);
+            snsBean.setInterval(interval);
+
+            // create RuleBean
+            RuleBean ruleBean = new RuleBean();
+            ruleBean.setName(name);
+            ruleBean.setDescription(description);
+            ruleBean.setData(data);
+            ruleBean.setCondition(condition);
+            ruleBean.setType("SNS");
+            ruleBean.setParentThing(ThingDAO.getInstance().get(parentThingId));
+
+
+            Set<ConstraintViolation<RuleBean>> constraintViolations = ValidatorService.getValidator().validate(ruleBean);
+
+            System.out.println(constraintViolations);
+
+            // create SnsAction in AWS
+            CreateTopicResult createTopicResult = SnsHelper.getInstance().createTopic(snsBean);
+            snsBean.setTopicARN(createTopicResult.getTopicArn());
+
+            try {
+                // create rule in AWS
+                CreateTopicRuleResult ruleResult = RuleHelper.getInstance().createTopicRule(ruleBean, snsBean);
+
+                // get the rule from AWS with for its ARN
+                String ruleArn = RuleHelper.getInstance().getTopicRule("thing" + parentThingId + "_sns_" + ruleBean.getName()).getRuleArn();
+
+                // add trigger permission in lambda function (notificationService) for this rule
+                AWSLambda lambda = AwsIotHelper.getAWSLambdaClient();
+                String functionArn = ConfigDAO.getInstance().get("lambdaNotificationArn").getValue();
+                lambda.addPermission(new AddPermissionRequest()
+                        .withFunctionName(functionArn)
+                        .withStatementId(UUID.randomUUID().toString())
+                        .withPrincipal("iot.amazonaws.com")
+                        .withSourceArn(ruleArn)
+                        .withAction("lambda:InvokeFunction"));
+
+                // add rule to DB
+                RuleDAO.getInstance().add(ruleBean);
+
+                // link rule and SNS in DB
+                snsBean.setParentRule(ruleBean);
+                SnsDAO.getInstance().add(snsBean);
+
+                // Get updated ruleBean
+                ruleBean = RuleDAO.getInstance().get(ruleBean.getId());
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return "{\"success\": false}";
+            }
+            return gson.toJson(ruleBean);
+        }
+        else throw new AccessDeniedException();
+    }
+
+
+    @GET
+    @Session
+    @Secure(roles = {RoleEnum.READ, RoleEnum.WRITE, RoleEnum.ALL})
+    @Path("/get/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String get(@PathParam("id") Integer ruleId) throws AccessDeniedException {
+        RuleBean ruleBean = RuleDAO.getInstance().get(ruleId);
+
+        ThingBean targetThing = ruleBean.getParentThing();
+        UserBean user = (UserBean)getSecurityContext().getUserPrincipal();
+
+        if (AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            return gson.toJson(ruleBean);
+        }
+        else throw new AccessDeniedException();
+    }
+
+
+    @PUT
+    @Path("/update/{id}")
+    @Session
+    @Secure(roles = {RoleEnum.ALL}, subjectType = "rule", subjectField = "parentId")
+    @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String update(@PathParam("id") Integer ruleId,
+                         @FormParam("description") String description,
+                         @FormParam("data") String data,
+                         @FormParam("condition") String condition,
+                         @FormParam("parentThing") Integer parentThingId) throws AccessDeniedException {
 
         // create RuleBean
-        RuleBean ruleBean = new RuleBean();
-        ruleBean.setName(name);
-        ruleBean.setDescription(description);
-        ruleBean.setData(data);
-        ruleBean.setCondition(condition);
-        ruleBean.setType("SNS");
-        ruleBean.setParentThing(ThingDAO.getInstance().get(parentThingId));
+        RuleBean ruleBean = RuleDAO.getInstance().get(ruleId);
 
+        ThingBean targetThing = ruleBean.getParentThing();
+        UserBean user = (UserBean)getSecurityContext().getUserPrincipal();
 
-        Set<ConstraintViolation<RuleBean>> constraintViolations = ValidatorService.getValidator().validate(ruleBean);
+        if (AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            ruleBean.setDescription(description);
+            ruleBean.setData(data);
+            ruleBean.setCondition(condition);
 
-        System.out.println(constraintViolations);
+            SnsBean snsBean = ruleBean.getSnsAction();
 
-        // create SnsAction in AWS
-        CreateTopicResult createTopicResult = SnsHelper.getInstance().createTopic(snsBean);
-        snsBean.setTopicARN(createTopicResult.getTopicArn());
-
-        try {
-            // create rule in AWS
-            CreateTopicRuleResult ruleResult = RuleHelper.getInstance().createTopicRule(ruleBean, snsBean);
-
-            // get the rule from AWS with for its ARN
-            String ruleArn = RuleHelper.getInstance().getTopicRule("thing" + parentThingId + "_sns_" + ruleBean.getName()).getRuleArn();
-
-            // add trigger permission in lambda function (notificationService) for this rule
-            AWSLambda lambda = AwsIotHelper.getAWSLambdaClient();
-            String functionArn = ConfigDAO.getInstance().get("lambdaNotificationArn").getValue();
-            lambda.addPermission(new AddPermissionRequest()
-                    .withFunctionName(functionArn)
-                    .withStatementId(UUID.randomUUID().toString())
-                    .withPrincipal("iot.amazonaws.com")
-                    .withSourceArn(ruleArn)
-                    .withAction("lambda:InvokeFunction"));
+            // update rule in AWS
+            RuleHelper.getInstance().replaceTopicRule(ruleBean, snsBean);
 
             // add rule to DB
-            RuleDAO.getInstance().add(ruleBean);
+            RuleDAO.getInstance().update(ruleBean);
 
             // link rule and SNS in DB
             snsBean.setParentRule(ruleBean);
@@ -109,66 +178,19 @@ public class SnsRuleResource extends BaseResource {
             // Get updated ruleBean
             ruleBean = RuleDAO.getInstance().get(ruleBean.getId());
 
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "{\"success\": false}";
+            return gson.toJson(ruleBean);
         }
-        return gson.toJson(ruleBean);
+        else throw new AccessDeniedException();
     }
 
-    @GET
-    @Session
-    @Path("/get/{id}")
-    @Secure(roles = {RoleEnum.ALL})
-    @Produces(MediaType.APPLICATION_JSON)
-    public String get(@PathParam("id") Integer ruleId) {
-        RuleBean ruleBean = RuleDAO.getInstance().get(ruleId);
-        return gson.toJson(ruleBean);
-    }
-
-    @PUT
-    @Session
-    @Path("/update/{id}")
-    @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-    @Produces(MediaType.APPLICATION_JSON)
-    @Secure(roles = {RoleEnum.ALL}, subjectType = "rule", subjectField = "parentId")
-    public String update(@PathParam("id") Integer ruleId,
-                         @FormParam("description") String description,
-                         @FormParam("data") String data,
-                         @FormParam("condition") String condition,
-                         @FormParam("parentThing") Integer parentThingId) {
-
-        // create RuleBean
-        RuleBean ruleBean = RuleDAO.getInstance().get(ruleId);
-        ruleBean.setDescription(description);
-        ruleBean.setData(data);
-        ruleBean.setCondition(condition);
-
-        SnsBean snsBean = ruleBean.getSnsAction();
-
-        // update rule in AWS
-        RuleHelper.getInstance().replaceTopicRule(ruleBean, snsBean);
-
-        // add rule to DB
-        RuleDAO.getInstance().update(ruleBean);
-
-        // link rule and SNS in DB
-        snsBean.setParentRule(ruleBean);
-        SnsDAO.getInstance().add(snsBean);
-
-        // Get updated ruleBean
-        ruleBean = RuleDAO.getInstance().get(ruleBean.getId());
-
-        return gson.toJson(ruleBean);
-    }
 
     /*TODO: Delete SNS topics when there is no rule in AWS pointing to it*/
     @DELETE
     @Path("/delete/{id}")
+    @Session
+    @Secure(roles = {RoleEnum.WRITE, RoleEnum.ALL})
     @Produces(MediaType.APPLICATION_JSON)
-    @Session()
-    @Secure(roles = {RoleEnum.ALL})
-    public String delete(@PathParam("id") Integer ruleId) {
+    public String delete(@PathParam("id") Integer ruleId) throws AccessDeniedException {
         /*
          * Steps:
          * 1. Get ruleBean
@@ -179,22 +201,29 @@ public class SnsRuleResource extends BaseResource {
         // get ruleBean
         RuleBean ruleBean = RuleDAO.getInstance().get(ruleId);
 
-        // delete rule in AWS
-        DeleteTopicRuleResult deleteTopicRuleResult = RuleHelper.getInstance().deleteRule(ruleBean);
+        ThingBean targetThing = ruleBean.getParentThing();
+        UserBean user = (UserBean)getSecurityContext().getUserPrincipal();
 
-        // delete rule bean which should also delete entries from SNS and SNSSubscriptions
-        RuleDAO.getInstance().delete(ruleId);
-        return "{\"success\": true}";
+        if (AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            // delete rule in AWS
+            DeleteTopicRuleResult deleteTopicRuleResult = RuleHelper.getInstance().deleteRule(ruleBean);
+
+            // delete rule bean which should also delete entries from SNS and SNSSubscriptions
+            RuleDAO.getInstance().delete(ruleId);
+            return "{\"success\": true}";
+        }
+        else throw new AccessDeniedException();
     }
 
+
     @DELETE
-    @Session
     @Path("/delete/")
+    @Session
+    @Secure(roles = {RoleEnum.WRITE, RoleEnum.ALL})
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.APPLICATION_JSON)
-    @Secure(roles = {RoleEnum.ALL})
     public String deleteByName(@NotNull @FormParam("name") String ruleName,
-                               @NotNull @FormParam("parentThing") Integer Id) {
+                               @NotNull @FormParam("parentThing") Integer Id) throws AccessDeniedException {
         /*
          * Steps:
          * 1. Get ruleBean
@@ -205,19 +234,25 @@ public class SnsRuleResource extends BaseResource {
         // get ruleBean
         RuleBean ruleBean = RuleDAO.getInstance().getByName(ruleName);
 
-        // delete rule in AWS
-        DeleteTopicRuleResult deleteTopicRuleResult = RuleHelper.getInstance().deleteRule(ruleBean);
+        ThingBean targetThing = ruleBean.getParentThing();
+        UserBean user = (UserBean)getSecurityContext().getUserPrincipal();
 
-        // delete rule bean which should also delete entries from SNS and SNSSubscriptions
-        RuleDAO.getInstance().deleteByName(ruleName);
-        return "{\"success\": true}";
+        if (AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            // delete rule in AWS
+            DeleteTopicRuleResult deleteTopicRuleResult = RuleHelper.getInstance().deleteRule(ruleBean);
+
+            // delete rule bean which should also delete entries from SNS and SNSSubscriptions
+            RuleDAO.getInstance().deleteByName(ruleName);
+            return "{\"success\": true}";
+        }
+        else throw new AccessDeniedException();
     }
 
 
     @POST
-    @Session
-    @Secure(roles = {RoleEnum.ALL})
     @Path("/subscribe/{id}")
+    @Session
+    @Secure(roles = {RoleEnum.WRITE, RoleEnum.ALL})
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.APPLICATION_JSON)
     public String subscribeSNSTopic(@PathParam("id") Integer snsId,
@@ -235,13 +270,20 @@ public class SnsRuleResource extends BaseResource {
         return gson.toJson(snsSubscriptionBean);
     }
 
+
     @GET
     @Path("/thing/{id}")
-    @Produces(MediaType.APPLICATION_JSON)
     @Session
-    @Secure(roles = {RoleEnum.ALL})
-    public String getByThing(@PathParam("id") Integer parentThingId) {
-        Set<RuleBean> ruleBean = RuleDAO.getInstance().getByThing(parentThingId);
-        return gson.toJson(ruleBean);
+    @Secure(roles = {RoleEnum.READ, RoleEnum.WRITE, RoleEnum.ALL})
+    @Produces(MediaType.APPLICATION_JSON)
+    public String getByThing(@PathParam("id") Integer parentThingId) throws AccessDeniedException {
+        ThingBean targetThing = ThingDAO.getInstance().get(parentThingId);
+        UserBean user = (UserBean)getSecurityContext().getUserPrincipal();
+
+        if(AuthorizationHelper.getInstance().checkAccess(user, targetThing)) {
+            Set<RuleBean> ruleBean = RuleDAO.getInstance().getByThing(parentThingId);
+            return gson.toJson(ruleBean);
+        }
+        else throw new AccessDeniedException();
     }
 }
